Updated Linux makefile, GTK# sample is now working.
- The Linux makefile is now generating the correct library name
(libssp_player.so instead of liblibssp_player.so); libssp_player is also
now looking in the current folder for BASS dynamic libraries instead of
only in the standard lib directories. This is what fixed the
DllNotFoundException in the GTK# sample.
- Updated GTK# sample; SSP_Init and SSP_InitDevice are now working fine,
but audio files haven't been tested yet.
- Added makefile/NetBeans project for player-sample-console. The sample
is working fine with audio files in Linux.