After several days of debugging, finally fixed barebones sample for Android with NVIDIA Tegra Nsight.
- Fixed JNI jstring handling with JNI UTF methods in ssp_java.c, this
seems the reason the app was crashing randomly on ARM devices when
allocing memory (malloc). Either Android drives me crazy or it looks
like the native stack stace didn't make sense because it complained
about something completely different than consuming the string... thanks
to NVIDIA Tegra Nsight I was able to pinpoint the problem. At least now
the sample works on Nexus 7 2012 and LG G2.
- Added NVIDIA Tegra Nsight project to /tegra. This is what enabled
proper debugging of the library. However the C code must be copied in
the directory in order for the project to work. I will try to revisit
that later to include the files outside the project directory. Using the
Android makefile doesn't make the C code debuggable, so the project has
to generate its own dynamic library instead of using libssp_player.